Add dependencies for SQLite3
Chose SQLite 3 for simplicity, speed, and portability over languages and
platforms.  Since partner is an international NGO using this app for a specific
purpose, high traffic is not important but languages and multiple platforms
are.  Assume these constraints will hold for similar organizations even at
large scale.  Downside is that SQLite may make it difficult to transition to
SQL later if required.

mattn/go-sqlite3 chosen because it is the most recommended sqlite driver for
Golang.  Golang requires a driver to run sqlite.

Selected links for further reading:
https://www.digitalocean.com/community/tutorials/how-and-when-to-use-sqlite
https://astaxie.gitbooks.io/build-web-application-with-golang/content/en/05.3.html