Add Error Page
Catch-all error page alerts user if message was not successfully sent to and
accepted by the Plivo API.

Previously, there was a bug where the "Message Sent" page was displayed
regardless of whether or not Plivo recieved the message.  This commit fixes
that problem.

There'll need to at least be a differentiation among error codes:
500: PocketAlert has had an internal problem, please let the developers know.
400: Something is missing/malformed in user's request.
401: User doesn't have the permission to send SMS (wrong Plivo auth).
404: The SMS network service (Plivo) couldn't be contacted.

Instead of creating a new web page for each scenario, it would be best to use
dynamic HTML.  Djano makes this easy, but this project uses Golang, not
Python.  Further investigation is needed to implement this feature.